Improve performance and consistency of alias lookups
The current alias lookup implementation adds a separate entry to the lookup map
for each IP address in the network of the alias. That's *a lot* of map entries
for, say, a ``/16`` network. This map becomes even larger if
``add_captid_to_resolve`` is enabled. This patch replaces this with a
treap-based lookup.

The improvement in performance depends on how many aliases you have and how
large your networks are. For our setup with ~80 aliases and a fair number of
``/19`` and ``/22`` networks, the speedup amounts to about 3x.

Alias lookups are now also more consistent. For aliases with partially
overlapping networks, the alias with the longest prefix match wins, not the one
that happens to be the last one to be added to the lookup map.

The new ``cidrtree`` dependency bumps the minimum Go version to 1.21.

Looks like it has been a while since the code base was formatted with gofmt, so
this patch unfortunately also includes some formatting changes for the changed
files.